Add ast module, move some stuff into files
